requirements.txt
pandas==2.2.2
pylint==3.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
optuna==3.6.1
matplotlib==3.10.1
numpy==2.0.1